pang0620 init
