+++ model/README.md
@@ -1,3 +1,36 @@
-# Intelproject5
-인텔프로젝트5팀
+# 인자 설명
 
+ref (필수, str)
+채점용 정답 영상 경로. 예: ref.mp4
+
+--cam (int, 기본 0)
+사용할 카메라 인덱스. 외장캠이면 보통 1 또는 2.
+
+--start (float, 기본 3.0)
+카운트다운 길이(초). 카운트다운 동안 정답 영상은 정지 프레임을 보여주고, 끝나면 0프레임부터 재생/채점 시작.
+
+--every (int, 기본 5)
+N프레임마다 채점. 예: --every 5면 5프레임에 한 번 포즈 비교/총점 업데이트. 크면 가벼움(빠름), 작으면 촘촘(정밀).
+
+--no-mirror (flag, 기본 꺼짐)
+기본은 거울 모드 ON(카메라 포즈 좌우 반전 후 채점). 이 플래그를 쓰면 거울 모드 비활성화.
+
+--disp-scale (float, 기본 1.0)
+표시 스케일. 1.0보다 작게 주면 윈도우 전체 축소. (저사양/큰 해상도에서 유용)
+
+--disp-width (int, 기본 없음)
+합성 화면의 가로 최대폭(px). 실제 스케일은 disp-scale과 이 값 중 더 보수적인 쪽이 적용됨.
+예: --disp-width 1280이면 창 가로가 1280px을 넘지 않도록 축소.
+
+--save (str, 기본 없음)
+합성(좌:REF, 우:USER) 출력 영상을 MP4로 저장할 경로. 예: --save out.mp4
+저장 FPS는 정답 영상 FPS를 따르고, 해상도는 (W*2, H).
+
+--imgsz (int, 기본 320)
+YOLO 입력 해상도. 작을수록 빠르고 크면 더 정확할 수 있음. (예: 256/320/416/512)
+
+--device (str, 기본 cuda 가능 시 'cuda', 아니면 'cpu')
+추론 디바이스. 예: --device cuda 또는 --device cpu.
+
+--half (flag, 기본 켜짐)
+FP16(Half precision) 사용. CUDA에서만 적용; CPU면 자동 무시. 속도 향상 기대.
